Moved inference results to prediction layer
